Refactoring - Extract class - Extracted event coordination helper methods from the PortletEventCoordinatationService into a new class named PortletEventCoordinationHelper.
