+++ uPortal-web/src/main/java/org/apereo/portal/portlet/rendering/PortletEventCoordinatationService.java
@@ -16,50 +16,20 @@ package org.apereo.portal.portlet.rendering;
 
 import com.google.common.base.Function;
 import com.google.common.collect.Lists;
-import java.io.Serializable;
-import java.io.StringReader;
-import java.util.Collections;
-import java.util.Iterator;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Queue;
-import java.util.Set;
+import java.util.*;
 import javax.portlet.Event;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Unmarshaller;
-import javax.xml.namespace.QName;
-import javax.xml.stream.XMLInputFactory;
-import javax.xml.stream.XMLStreamException;
-import javax.xml.stream.XMLStreamReader;
 import net.sf.ehcache.Ehcache;
-import net.sf.ehcache.Element;
-import org.apache.commons.lang.StringUtils;
 import org.apache.pluto.container.PortletContainer;
-import org.apache.pluto.container.PortletContainerException;
 import org.apache.pluto.container.PortletWindow;
 import org.apache.pluto.container.driver.PortletContextService;
-import org.apache.pluto.container.om.portlet.ContainerRuntimeOption;
-import org.apache.pluto.container.om.portlet.EventDefinition;
 import org.apache.pluto.container.om.portlet.EventDefinitionReference;
-import org.apache.pluto.container.om.portlet.PortletApplicationDefinition;
 import org.apache.pluto.container.om.portlet.PortletDefinition;
 import org.apereo.portal.EntityIdentifier;
 import org.apereo.portal.IUserPreferencesManager;
 import org.apereo.portal.layout.IUserLayoutManager;
-import org.apereo.portal.portlet.container.EventImpl;
-import org.apereo.portal.portlet.om.IPortletDefinition;
-import org.apereo.portal.portlet.om.IPortletDefinitionId;
-import org.apereo.portal.portlet.om.IPortletDefinitionParameter;
-import org.apereo.portal.portlet.om.IPortletEntity;
-import org.apereo.portal.portlet.om.IPortletEntityId;
-import org.apereo.portal.portlet.om.IPortletWindow;
-import org.apereo.portal.portlet.om.IPortletWindowId;
+import org.apereo.portal.portlet.om.*;
 import org.apereo.portal.portlet.registry.IPortletDefinitionRegistry;
 import org.apereo.portal.portlet.registry.IPortletEntityRegistry;
 import org.apereo.portal.portlet.registry.IPortletWindowRegistry;
@@ -69,7 +39,6 @@ import org.apereo.portal.services.AuthorizationServiceFacade;
 import org.apereo.portal.url.IPortalRequestUtils;
 import org.apereo.portal.user.IUserInstance;
 import org.apereo.portal.user.IUserInstanceManager;
-import org.apereo.portal.utils.Tuple;
 import org.apereo.portal.utils.web.PortalWebUtils;
 import org.apereo.portal.xml.XmlUtilities;
 import org.slf4j.Logger;
@@ -89,8 +58,6 @@ import org.springframework.stereotype.Service;
  */
 @Service("eventCoordinationService")
 public class PortletEventCoordinatationService implements IPortletEventCoordinationService {
-
-    public static final String GLOBAL_EVENT__CONTAINER_OPTION = "org.apereo.portal.globalEvent";
 
     private static final String PORTLET_EVENT_QUEUE =
             PortletEventCoordinatationService.class.getName() + ".PORTLET_EVENT_QUEUE";
@@ -147,6 +114,14 @@ public class PortletEventCoordinatationService implements IPortletEventCoordinat
     public void setPortletEntityRegistry(IPortletEntityRegistry portletEntityRegistry) {
         this.portletEntityRegistry = portletEntityRegistry;
     }
+
+    private final PortletEventCoordinationHelper portletEventCoordinationHelper =
+            new PortletEventCoordinationHelper(
+                    this.xmlUtilities,
+                    this.portletContextService,
+                    this.portletWindowRegistry,
+                    this.supportedEventCache,
+                    this.portletDefinitionRegistry);
 
     /**
      * Returns a request scoped PortletEventQueue used to track events to process and events to
@@ -230,7 +205,9 @@ public class PortletEventCoordinatationService implements IPortletEventCoordinat
             final IPortletWindowId sourceWindowId = queuedEvent.getPortletWindowId();
             final Event event = queuedEvent.getEvent();
 
-            final boolean globalEvent = isGlobalEvent(request, sourceWindowId, event);
+            final boolean globalEvent =
+                    this.portletEventCoordinationHelper.isGlobalEvent(
+                            request, sourceWindowId, event);
 
             final Set<IPortletDefinition> portletDefinitions =
                     new LinkedHashSet<IPortletDefinition>();
@@ -285,7 +262,7 @@ public class PortletEventCoordinatationService implements IPortletEventCoordinat
                 final IPortletDefinition portletDefinition = portletEntity.getPortletDefinition();
                 final IPortletDefinitionId portletDefinitionId =
                         portletDefinition.getPortletDefinitionId();
-                if (this.supportsEvent(event, portletDefinitionId)) {
+                if (portletEventCoordinationHelper.supportsEvent(event, portletDefinitionId)) {
                     this.logger.debug("{} supports event {}", portletDefinition, event);
 
                     // If this is the default portlet entity remove the definition from the all defs
@@ -318,7 +295,8 @@ public class PortletEventCoordinatationService implements IPortletEventCoordinat
                     for (final IPortletWindow portletWindow : portletWindows) {
                         this.logger.debug("{} resolved target {}", event, portletWindow);
                         final IPortletWindowId portletWindowId = portletWindow.getPortletWindowId();
-                        final Event unmarshalledEvent = this.unmarshall(portletWindow, event);
+                        final Event unmarshalledEvent =
+                                portletEventCoordinationHelper.unmarshall(portletWindow, event);
                         portletEventQueue.offerEvent(
                                 portletWindowId,
                                 new QueuedEvent(sourceWindowId, unmarshalledEvent));
@@ -356,7 +334,8 @@ public class PortletEventCoordinatationService implements IPortletEventCoordinat
                             portletDefinition.getPortletDefinitionId();
                     // Check if the user can render the portlet definition before doing event tests
                     if (ap.canRender(portletDefinitionId.getStringId())) {
-                        if (this.supportsEvent(event, portletDefinitionId)) {
+                        if (portletEventCoordinationHelper.supportsEvent(
+                                event, portletDefinitionId)) {
                             this.logger.debug("{} supports event {}", portletDefinition, event);
 
                             final IPortletEntity portletEntity =
@@ -373,7 +352,8 @@ public class PortletEventCoordinatationService implements IPortletEventCoordinat
                                 final IPortletWindowId portletWindowId =
                                         portletWindow.getPortletWindowId();
                                 final Event unmarshalledEvent =
-                                        this.unmarshall(portletWindow, event);
+                                        portletEventCoordinationHelper.unmarshall(
+                                                portletWindow, event);
                                 portletEventQueue.offerEvent(
                                         portletWindowId,
                                         new QueuedEvent(sourceWindowId, unmarshalledEvent));
@@ -383,198 +363,5 @@ public class PortletEventCoordinatationService implements IPortletEventCoordinat
                 }
             }
         }
-    }
-
-    protected boolean isGlobalEvent(
-            HttpServletRequest request, IPortletWindowId sourceWindowId, Event event) {
-        final IPortletWindow portletWindow =
-                this.portletWindowRegistry.getPortletWindow(request, sourceWindowId);
-        final IPortletEntity portletEntity = portletWindow.getPortletEntity();
-        final IPortletDefinition portletDefinition = portletEntity.getPortletDefinition();
-        final IPortletDefinitionId portletDefinitionId = portletDefinition.getPortletDefinitionId();
-        final PortletApplicationDefinition parentPortletApplicationDescriptor =
-                this.portletDefinitionRegistry.getParentPortletApplicationDescriptor(
-                        portletDefinitionId);
-
-        final ContainerRuntimeOption globalEvents =
-                parentPortletApplicationDescriptor.getContainerRuntimeOption(
-                        GLOBAL_EVENT__CONTAINER_OPTION);
-        if (globalEvents != null) {
-            final QName qName = event.getQName();
-            final String qNameStr = qName.toString();
-            for (final String globalEvent : globalEvents.getValues()) {
-                if (qNameStr.equals(globalEvent)) {
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    protected Event unmarshall(IPortletWindow portletWindow, Event event) {
-        // TODO make two types of Event impls, one for marshalled data and one for unmarshalled data
-        String value = (String) event.getValue();
-
-        final XMLInputFactory xmlInputFactory = this.xmlUtilities.getXmlInputFactory();
-        final XMLStreamReader xml;
-        try {
-            xml = xmlInputFactory.createXMLStreamReader(new StringReader(value));
-        } catch (XMLStreamException e) {
-            throw new IllegalStateException(
-                    "Failed to create XMLStreamReader for portlet event: " + event, e);
-        }
-
-        // now test if object is jaxb
-        final EventDefinition eventDefinitionDD =
-                getEventDefinition(portletWindow, event.getQName());
-
-        final PortletDefinition portletDefinition =
-                portletWindow.getPlutoPortletWindow().getPortletDefinition();
-        final PortletApplicationDefinition application = portletDefinition.getApplication();
-        final String portletApplicationName = application.getName();
-
-        final ClassLoader loader;
-        try {
-            loader = portletContextService.getClassLoader(portletApplicationName);
-        } catch (PortletContainerException e) {
-            throw new IllegalStateException(
-                    "Failed to get ClassLoader for portlet application: " + portletApplicationName,
-                    e);
-        }
-
-        final String eventType = eventDefinitionDD.getValueType();
-        final Class<? extends Serializable> clazz;
-        try {
-            clazz = loader.loadClass(eventType).asSubclass(Serializable.class);
-        } catch (ClassNotFoundException e) {
-            throw new IllegalArgumentException(
-                    "Declared event type '"
-                            + eventType
-                            + "' cannot be found in portlet application: "
-                            + portletApplicationName,
-                    e);
-        }
-
-        // TODO cache JAXBContext in registered portlet application
-        final JAXBElement<? extends Serializable> result;
-        try {
-            final JAXBContext jc = JAXBContext.newInstance(clazz);
-            final Unmarshaller unmarshaller = jc.createUnmarshaller();
-            result = unmarshaller.unmarshal(xml, clazz);
-        } catch (JAXBException e) {
-            throw new IllegalArgumentException(
-                    "Cannot create JAXBContext for event type '"
-                            + eventType
-                            + "' from portlet application: "
-                            + portletApplicationName,
-                    e);
-        }
-
-        return new EventImpl(event.getQName(), result.getValue());
-    }
-
-    // TODO cache this resolution
-    protected EventDefinition getEventDefinition(IPortletWindow portletWindow, QName name) {
-        PortletApplicationDefinition appDD =
-                portletWindow.getPlutoPortletWindow().getPortletDefinition().getApplication();
-        for (EventDefinition def : appDD.getEventDefinitions()) {
-            if (def.getQName() != null) {
-                if (def.getQName().equals(name)) return def;
-            } else {
-                QName tmp = new QName(appDD.getDefaultNamespace(), def.getName());
-                if (tmp.equals(name)) return def;
-            }
-        }
-        throw new IllegalStateException();
-    }
-
-    protected Set<QName> getAllAliases(
-            QName eventName, PortletApplicationDefinition portletApplicationDefinition) {
-        final List<? extends EventDefinition> eventDefinitions =
-                portletApplicationDefinition.getEventDefinitions();
-        if (eventDefinitions == null || eventDefinitions.isEmpty()) {
-            return Collections.emptySet();
-        }
-
-        final String defaultNamespace = portletApplicationDefinition.getDefaultNamespace();
-
-        for (final EventDefinition eventDefinition : eventDefinitions) {
-            final QName defQName = eventDefinition.getQualifiedName(defaultNamespace);
-            if (defQName != null && defQName.equals(eventName)) {
-                final List<QName> aliases = eventDefinition.getAliases();
-                if (aliases == null || aliases.isEmpty()) {
-                    return Collections.emptySet();
-                }
-
-                return new LinkedHashSet<QName>(aliases);
-            }
-        }
-
-        return Collections.emptySet();
-    }
-
-    protected boolean supportsEvent(Event event, IPortletDefinitionId portletDefinitionId) {
-        final QName eventName = event.getQName();
-
-        // The cache key to use
-        final Tuple<IPortletDefinitionId, QName> key =
-                new Tuple<IPortletDefinitionId, QName>(portletDefinitionId, eventName);
-
-        // Check in the cache if the portlet definition supports this event
-        final Element element = this.supportedEventCache.get(key);
-        if (element != null) {
-            final Boolean supported = (Boolean) element.getObjectValue();
-            if (supported != null) {
-                return supported;
-            }
-        }
-
-        final PortletApplicationDefinition portletApplicationDescriptor =
-                this.portletDefinitionRegistry.getParentPortletApplicationDescriptor(
-                        portletDefinitionId);
-        if (portletApplicationDescriptor == null) {
-            return false;
-        }
-
-        final Set<QName> aliases = this.getAllAliases(eventName, portletApplicationDescriptor);
-
-        final String defaultNamespace = portletApplicationDescriptor.getDefaultNamespace();
-
-        // No support found so far, do more complex namespace matching
-        final PortletDefinition portletDescriptor =
-                this.portletDefinitionRegistry.getParentPortletDescriptor(portletDefinitionId);
-        if (portletDescriptor == null) {
-            return false;
-        }
-
-        final List<? extends EventDefinitionReference> supportedProcessingEvents =
-                portletDescriptor.getSupportedProcessingEvents();
-        for (final EventDefinitionReference eventDefinitionReference : supportedProcessingEvents) {
-            final QName qualifiedName = eventDefinitionReference.getQualifiedName(defaultNamespace);
-            if (qualifiedName == null) {
-                continue;
-            }
-
-            // See if the supported qname and event qname match explicitly
-            // Look for alias names
-            if (qualifiedName.equals(eventName) || aliases.contains(qualifiedName)) {
-                this.supportedEventCache.put(new Element(key, Boolean.TRUE));
-                return true;
-            }
-
-            // Look for namespaced events
-            if (StringUtils.isEmpty(qualifiedName.getNamespaceURI())) {
-                final QName namespacedName =
-                        new QName(defaultNamespace, qualifiedName.getLocalPart());
-                if (eventName.equals(namespacedName)) {
-                    this.supportedEventCache.put(new Element(key, Boolean.TRUE));
-                    return true;
-                }
-            }
-        }
-
-        this.supportedEventCache.put(new Element(key, Boolean.FALSE));
-        return false;
     }
 }
